💾 消息入 SQLite：聊天记录跨重启不丢
删掉内存环形队列（chat Vec + chat_seq），push_chat 直写 DB
（自增主键即消息 id，出站图片缓存 key 因此跨重启稳定），
新增 set_msg_status 供状态流转；面板轮询改读 recent_msgs(200)，
响应附带 status / srcPath。启动时残留的 sending 统一落 fail，
日志提示可手动重试。

+++ rust/src/db.rs
@@ -109,7 +109,7 @@ impl Db {
     pub fn fail_pending_sending(&self) -> usize {
         let Ok(c) = self.0.lock() else { return 0 };
         c.execute("UPDATE messages SET status='fail' WHERE out=1 AND status='sending'", [])
-            .unwrap_or(0) as usize
+            .unwrap_or(0)
     }
 
     /// 启动时裁剪历史（只留最近 MSG_KEEP 条）
